All working ok with MRHOF2
One minor change (if of== mrhof2) in rpl-dag.c to support the big change in parent's weight.
the mrhof2 starts altering the path at the sender, and ends the alteration at the receiver.
There is one major problem:  the ()RED) node(s) within the altered path have two red parents. They choose the best parent based on etx, which is not always the case....
Ideally, the sender should be number x, and the receiver No 0. Each red node along the path, should be x-1, x-2, until 0. Hence, each red node, if it has two parents, it chooses, the red parent with smaller x.
But this requires another variable....